+++ scripts/rnaseq_shinyhelper_functions.R
@@ -234,11 +234,10 @@ manhattan_plot = function(res_dt=gwFRASER,sample = 'HSJ_036_03_PAX',top=25,pcuto
 }
 
 
-#
 ###
-###Genereate a gene prioritisation data.frame based on gene lists (HPO, outrider, fraser) 
+###Generate a gene prioritisation data.frame based on gene lists (HPO, outrider, fraser) 
 ###
-gene_prioritization = function(sample = 'HSJ_001_03_PAX',fraser_padjust=0.05,outrider_pvalue=0.01,top=100,hpo_sample=clinical,hpo_all=file.path(params$datadir,'genes_to_phenotype.txt'),fraser="",outrider="",custom_genes=""){
+gene_prioritization = function(sample = 'HSJ_001_03_PAX',top=100,hpo_sample=clinical,hpo_all=file.path(params$datadir,'genes_to_phenotype.txt'),fraser="",outrider="",custom_genes=""){
   
   #hpo 
   if(!file.exists(hpo_all)) {download.file(url='https://github.com/obophenotype/human-phenotype-ontology/releases/latest/download/genes_to_phenotype.txt',dest=file.path(hpo_all))} else {print(paste0('file ', hpo_all,'exists'))}
@@ -258,29 +257,36 @@ gene_prioritization = function(sample = 'HSJ_001_03_PAX',fraser_padjust=0.05,out
   
   #add outrider
   if(!is.null(dim(outrider))) {
-    hpo_genes$outrider = outrider$geneID[outrider$sampleID==sample & outrider$pValue < outrider_pvalue]
-    hpo_genes$outrider = hpo_genes$outrider[!is.na(hpo_genes$outrider)]    
+   outrider_temp = outrider[outrider$sampleID==sample,colnames(outrider) %in% c('geneID','pValue','zScore','exon_zScore','exon_pValue')]
+   outrider_temp = outrider_temp[!is.na(outrider_temp$geneID),]    
+   colnames(outrider_temp) = c('geneID','OUTRIDER_pValue','OUTRIDER_zScore','OUTRIDER_exon_pValue','OUTRIDER_exon_zScore')
   }
   
   #add fraser
   if(!is.null(dim(fraser))) {
-    hpo_genes$fraser = fraser$hgncSymbol[fraser$sampleID==sample & fraser$padjust < fraser_padjust]
-    hpo_genes$fraser = hpo_genes$fraser[!is.na(hpo_genes$fraser)]    
+    fraser_temp = fraser[fraser$sampleID==sample,colnames(fraser) %in% c('hgncSymbol','pValue')]
+    fraser_temp = fraser_temp[!is.na(fraser_temp$hgncSymbol),]    
+    colnames(fraser_temp) = c('geneID','FRASER_pValue')
+    fraser_temp = fraser_temp[!duplicated(fraser_temp$geneID),]
   }
   
-  #add custom
-  if(!is.null(dim(custom_genes))) {
-    hpo_genes$custom_genes = custom_genes[,1]
-  }
-  
+  #merge Outliers
+  outlier_temp = merge(fraser_temp,outrider_temp,by= 'geneID',all=T, sort=F)
+   
   #generate a big table stating which gene is listed where.
   all_genes = unique(unlist(hpo_genes))
   table = sapply(hpo_genes, function(x) all_genes %in% x)
-  table = as.data.frame(table)
-  rownames(table) = all_genes
-  table = data.frame(gene_score = rowSums(table),table, check.names = F)
+  table = data.frame(gene_score = rowSums(table),table*1, check.names = F)
+  table$geneID = all_genes
+  table = merge(table,outlier_temp,by= 'geneID',all.x= T, sort=F)
+  table$OUTRIDER_pValue[is.na(table$OUTRIDER_pValue)] = 'ns'
+  table$OUTRIDER_zScore[is.na(table$OUTRIDER_zScore)] = 'ns'
+  table$OUTRIDER_exon_pValue[is.na(table$OUTRIDER_exon_pValue)] = 'ns'
+  table$OUTRIDER_exon_zScore[is.na(table$OUTRIDER_exon_zScore)] = 'ns'
+  table$FRASER_padjust[is.na(table$FRASER_padjust)] = 'ns'
+  table = table[,c(1:2,ncol(table):3)]
   table = table[order(table$gene_score,decreasing = T),]
-  table = table[,c(1,ncol(table):2)]*1
+ 
 
   #return top hist
   return(head(table,top))